Fixed issue #9328

+++ packages/Webkul/Theme/src/Repositories/ThemeCustomizationRepository.php
@@ -20,23 +20,54 @@ class ThemeCustomizationRepository extends Repository
     }
 
     /**
-     * Upload images
+     * Update the specified theme
      *
-     * @param  array  $imageOptions
-     * @param  \Webkul\Shop\Contracts\ThemeCustomization  $theme
-     * @return void
+     * @param  array  $data
+     * @param  int  $id
      */
-    public function uploadImage($imageOptions, $theme, $deletedSliderImages = [])
+    public function update($data, $id): ThemeCustomization
     {
-        foreach ($deletedSliderImages as $slider) {
-            Storage::delete(str_replace('storage/', '', $slider['image']));
+        $locale = core()->getRequestedLocaleCode();
+
+        if ($data['type'] == 'static_content') {
+            $data[$locale]['options']['html'] = preg_replace('/<script\b[^>]*>(.*?)<\/script>/is', '', $data[$locale]['options']['html']);
+            $data[$locale]['options']['css'] = preg_replace('/<script\b[^>]*>(.*?)<\/script>/is', '', $data[$locale]['options']['css']);
         }
 
-        if (isset($imageOptions['options'])) {
+        if (in_array($data['type'], ['image_carousel', 'services_content'])) {
+            unset($data[$locale]['options']);
+        }
+
+        $theme = parent::update($data, $id);
+
+        if (in_array($data['type'], ['image_carousel', 'services_content'])) {
+            $this->uploadImage(request()->all(), $theme);
+        }
+
+        return $theme;
+    }
+
+    /**
+     * Upload images
+     *
+     * @param array data
+     * @param  \Webkul\Theme\Contracts\ThemeCustomization  $name
+     * @return void|string
+     */
+    public function uploadImage(array $data, $theme)
+    {
+        $locale = core()->getRequestedLocaleCode();
+
+        if (isset($data[$locale]['deleted_sliders'])) {
+            foreach ($data[$locale]['deleted_sliders'] as $slider) {
+                Storage::delete(str_replace('storage/', '', $slider['image']));
+            }
+        }
+
+        if (isset($data[$locale]['options'])) {
             $options = [];
 
-            foreach ($imageOptions['options'] as $image) {
-
+            foreach ($data[$locale]['options'] as $image) {
                 if (isset($image['service_icon'])) {
                     $options['services'][] = [
                         'service_icon' => $image['service_icon'],
@@ -52,8 +83,8 @@ class ThemeCustomizationRepository extends Repository
                     Storage::put($path, $manager->make($image['image'])->encode('webp'));
 
                     if (
-                        isset($imageOptions['type'])
-                        && $imageOptions['type'] == 'static_content'
+                        isset($data['type'])
+                        && $data['type'] == 'static_content'
                     ) {
                         return Storage::url($path);
                     }
@@ -67,10 +98,9 @@ class ThemeCustomizationRepository extends Repository
                 }
             }
 
-            $translatedModel = $theme->translate(core()->getRequestedLocaleCode());
-
+            $translatedModel = $theme->translate($locale);
             $translatedModel->options = $options ?? [];
-
+            $translatedModel->theme_customization_id = $theme->id;
             $translatedModel->save();
         }
     }
